// frontend/src/api/schema.ts
export interface components {
    schemas: {
// ...
        GuestBookingConfirmation: {
            id: string;
            eventTypeId: string;
            /** Format: date-time */
            startTime: string;
            /** Format: date-time */
            endTime: string;
            guestName: string;
            guestEmail: string;
            notes?: string;
            /** Format: date-time */
            createdAt: string;
            eventTypeTitle: string;
            /** Format: int32 */
            durationInMinutes: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}